Cleanup bin/console and bin/setup files
This comes out of the box when you generate a new gem and we're not actually wiring them up anywhere. Once we remove these we can also remove more gemfile/gemspec files since we can assume there's nothing in the gem being appbundled. We're doing this in Chef 16:

Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/strings-ansi-0.2.0/bin/console
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/strings-ansi-0.2.0/bin/setup
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/power_assert-1.1.7/bin/console
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/power_assert-1.1.7/bin/setup
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/rake-12.3.3/bin/console
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/rake-12.3.3/bin/setup
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/wisper-2.0.1/bin/console
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/wisper-2.0.1/bin/setup
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/mini_mime-1.0.2/bin/console
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/mini_mime-1.0.2/bin/setup
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/net-telnet-0.2.0/bin/console
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/net-telnet-0.2.0/bin/setup
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/xmlrpc-0.3.0/bin/console
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/xmlrpc-0.3.0/bin/setup
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/rubocop-0.83.0/bin/console
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/rubocop-0.83.0/bin/setup
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/sync-0.5.0/bin/console
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/sync-0.5.0/bin/setup
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/json_schemer-0.2.11/bin/console
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/json_schemer-0.2.11/bin/setup
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/rake-13.0.1/bin/console
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/rake-13.0.1/bin/setup
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/tty-reader-0.7.0/bin/console
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/tty-reader-0.7.0/bin/setup
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/tty-box-0.5.0/bin/console
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/tty-box-0.5.0/bin/setup
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/strings-0.1.8/bin/console
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/strings-0.1.8/bin/setup
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/public_suffix-4.0.5/bin/console
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/fog-openstack-1.0.10/bin/console
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/fog-openstack-1.0.10/bin/setup
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/strings-ansi-0.2.0/strings-ansi.gemspec
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/strings-ansi-0.2.0/Rakefile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/strings-ansi-0.2.0/tasks
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/power_assert-1.1.7/power_assert.gemspec
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/power_assert-1.1.7/Gemfile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/power_assert-1.1.7/Rakefile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/wisper-2.0.1/wisper.gemspec
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/wisper-2.0.1/Gemfile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/wisper-2.0.1/Rakefile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/mini_mime-1.0.2/mini_mime.gemspec
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/mini_mime-1.0.2/Gemfile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/mini_mime-1.0.2/Rakefile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/net-telnet-0.2.0/net-telnet.gemspec
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/net-telnet-0.2.0/Gemfile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/net-telnet-0.2.0/Rakefile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/xmlrpc-0.3.0/xmlrpc.gemspec
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/xmlrpc-0.3.0/Gemfile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/xmlrpc-0.3.0/Rakefile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/sync-0.5.0/sync.gemspec
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/sync-0.5.0/Gemfile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/sync-0.5.0/Rakefile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/json_schemer-0.2.11/json_schemer.gemspec
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/json_schemer-0.2.11/Gemfile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/json_schemer-0.2.11/Rakefile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/tty-reader-0.7.0/tty-reader.gemspec
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/tty-reader-0.7.0/Rakefile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/tty-reader-0.7.0/tasks
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/tty-box-0.5.0/tty-box.gemspec
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/tty-box-0.5.0/Rakefile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/tty-box-0.5.0/tasks
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/strings-0.1.8/strings.gemspec
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/strings-0.1.8/Rakefile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/strings-0.1.8/tasks
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/public_suffix-4.0.5/public_suffix.gemspec
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/public_suffix-4.0.5/Gemfile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/public_suffix-4.0.5/Rakefile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/fog-openstack-1.0.10/fog-openstack.gemspec
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/fog-openstack-1.0.10/Gemfile
Deleting /opt/chef-workstation/embedded/lib/ruby/gems/2.7.0/gems/fog-openstack-1.0.10/Rakefile

+++ omnibus/config/software/more-ruby-cleanup.rb
@@ -27,6 +27,13 @@ source path: "#{project.files_path}/#{name}"
 dependency "ruby"
 
 build do
+  block "Removing console and setup binaries" do
+    Dir.glob("#{install_dir}/embedded/lib/ruby/gems/*/gems/*/bin/{console,setup}").each do |f|
+      puts "Deleting #{f}"
+      FileUtils.rm_rf(f)
+    end
+  end
+
   block "remove any .gitkeep files" do
     Dir.glob("#{install_dir}/**/{.gitkeep,.keep}").each do |f|
       puts "Deleting #{f}"
@@ -98,8 +105,8 @@ build do
       }
 
       Dir.glob(Dir.glob("#{target_dir}/*/{#{files.join(",")}}")).each do |f|
-        # don't delete these files if there's a bin dir in the same dir
-        unless Dir.exist?(File.join(File.dirname(f), "bin"))
+        # don't delete these files if there's a non-empty bin dir in the same dir
+        unless Dir.exist?(File.join(File.dirname(f), "bin")) && !Dir.empty?(File.join(File.dirname(f), "bin"))
           puts "Deleting #{f}"
           FileUtils.rm_rf(f)
         end
